# docs/openapi-admin.yaml
openapi: 3.1.0
info:
  title: Wallet Backend Admin API
  description: |
    Internal administration API for managing the wallet backend multi-tenant infrastructure.
    
    This API is intended to be exposed on a separate port (default: 8081) and should only be
    accessible from internal networks or trusted management systems.
    
    ## Multi-Tenancy Model
    
    The wallet backend supports multiple tenants, each with their own isolated:
    - Users and credentials
    - Credential issuers (where users can request credentials)
    - Verifiers (where users can present credentials)
    
    Users are associated with tenants through memberships. Each user can belong to multiple
    tenants with different roles (user, admin).
    
    ## OIDC Gate
    
    Tenants can optionally require users to authenticate with an enterprise Identity Provider
    (OpenID Connect) before accessing registration and/or login endpoints. The gate supports:
    - `registration` mode: OIDC required for new wallet registrations
    - `login` mode: OIDC required for wallet logins
    - `both` mode: OIDC required for registration and login (with separate OPs per operation)
    - `none` mode: default open behavior (no OIDC pre-authentication)
    
    When enabled, a separate `registration_op` and/or `login_op` (OIDC Provider Config) must
    be supplied. An optional `bind_identity` flag permanently links the enterprise IdP subject
...
    
    ## Verifier Client ID
    
...
    
    ## Security
    
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /admin/tenants/{tenantId}/instances:
    get:
      tags:
        - Wallet Instances
      summary: List wallet instances
      description: Returns all wallet instances registered under this tenant.
      operationId: listWalletInstances
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: List of wallet instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WalletInstance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/tenants/{tenantId}/instances/{instanceId}:
    get:
      tags:
        - Wallet Instances
      summary: Get a wallet instance
      description: Returns details of a specific wallet instance by its JWK thumbprint ID.
      operationId: getWalletInstance
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/instanceId'
      responses:
        '200':
          description: Wallet instance details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletInstance'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Wallet Instances
      summary: Delete a wallet instance
      description: Hard-deletes a wallet instance. Emits an audit event.
      operationId: deleteWalletInstance
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/instanceId'
      responses:
        '204':
          description: Instance deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/tenants/{tenantId}/instances/{instanceId}/status:
    put:
      tags:
        - Wallet Instances
      summary: Update wallet instance status
      description: |
        Changes the status of a wallet instance (active, suspended, or revoked).
        Emits an audit event for the status transition.
      operationId: updateWalletInstanceStatus
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/instanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [active, suspended, revoked]
                reason:
                  type: string
                  description: Optional reason for the status change
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/tenants/{tenantId}/users/{userId}/instances:
    get:
      tags:
        - Wallet Instances
      summary: List wallet instances by user
      description: Returns all wallet instances belonging to a specific user within a tenant.
      operationId: listWalletInstancesByUser
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: List of wallet instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WalletInstance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/tenants/{tenantId}/users/{userId}/detail:
    get:
      tags:
        - Users
      summary: Get user detail
      description: |
        Returns extended user details including passkey information, DID, and wallet type.
        Only non-PII fields are exposed.
      operationId: getUserDetail
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  uuid:
                    type: string
                  did:
                    type: string
                  wallet_type:
                    type: string
                  passkeys:
                    type: array
                    items:
                      type: object
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/tenants/{tenantId}/stats:
    get:
      tags:
        - Tenants
      summary: Get tenant statistics
      description: Returns aggregated statistics for a tenant. Not yet implemented.
      operationId: getTenantStats
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '501':
          description: Not yet implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      description: The unique identifier of the tenant
      schema:
        type: string
        pattern: '^[a-z][a-z0-9-]{1,49}$'
      example: acme-corp
    userId:
      name: userId
      in: path
      required: true
      description: The UUID of the user
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    instanceId:
      name: instanceId
      in: path
      required: true
      description: The JWK thumbprint identifier of the wallet instance
      schema:
        type: string
      example: "dGVzdC10aHVtYnByaW50"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      example:
        error: "Resource not found"
...
    WalletInstance:
      type: object
      properties:
        id:
          type: string
          description: JWK thumbprint identifier
        tenant_id:
          type: string
        user_id:
          type: string
        status:
          type: string
          enum: [active, suspended, revoked]
        wscd_type:
          type: string
          enum: [web_crypto, remote, native_ios, native_android]
        credential_id:
          type: string
        r2ps_client_id:
          type: string
        device_info:
          type: object
          properties:
            platform:
              type: string
            os:
              type: string
            model:
              type: string
            app_id:
              type: string
        attestation_source:
          type: string
        last_attested_at:
          type: string
          format: date-time
        attestation_count:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deactivated_at:
          type: string
          format: date-time
        deactivation_reason:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Tenant not found"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Failed to process request"
    Unauthorized:
      description: Missing or invalid admin bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Invalid token"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static bearer token for admin API authentication. Configure via the
        `WALLET_SERVER_ADMIN_TOKEN` environment variable or `server.admin_token`
        config key. If not set, a random token is generated at startup and
        logged (development mode only).